fix(bridge): display deposit tx link in pending state (#5942)
