+++ apps/cowswap-frontend/src/modules/bridge/pure/contents/BridgingProgressContent/ReceivedBridgingContent/index.tsx
@@ -1,19 +1,18 @@
 import { ReactNode } from 'react'
 
-import ReceiptIcon from '@cowprotocol/assets/cow-swap/icon-receipt.svg'
 import { getChainInfo } from '@cowprotocol/common-const'
-import { useMediaQuery } from '@cowprotocol/common-hooks'
 import { ExplorerDataType, getExplorerLink } from '@cowprotocol/common-utils'
 import { BridgeStatusResult, SupportedChainId } from '@cowprotocol/cow-sdk'
-import { ExternalLink, Media } from '@cowprotocol/ui'
 import { Currency, CurrencyAmount, Token } from '@uniswap/sdk-core'
 
 import { useBridgeSupportedNetworks } from 'entities/bridgeProvider'
 
 import { ConfirmDetailsItem, ReceiveAmountTitle } from 'modules/trade'
 
-import { StyledTimelineReceiptIcon, SuccessTextBold, TimelineIconCircleWrapper } from '../../../../styles'
+import { SuccessTextBold } from '../../../../styles'
+import { DepositTxLink } from '../../../DepositTxLink'
 import { TokenAmountDisplay } from '../../../TokenAmountDisplay'
+import { TransactionLinkItem } from '../../../TransactionLink'
 
 interface ReceivedBridgingContentProps {
   statusResult?: BridgeStatusResult
@@ -21,35 +20,6 @@ interface ReceivedBridgingContentProps {
   destinationChainId: number
   receivedAmount: CurrencyAmount<Currency>
   receivedAmountUsd: CurrencyAmount<Token> | null | undefined
-}
-
-interface TransactionLinkProps {
-  link: string
-  label: string
-  chainId: number
-}
-
-function TransactionLink({ link, label, chainId }: TransactionLinkProps): ReactNode {
-  const { data: bridgeSupportedNetworks } = useBridgeSupportedNetworks()
-  const bridgeNetwork = bridgeSupportedNetworks?.find((network) => network.id === chainId)
-  const isMobile = useMediaQuery(Media.upToSmall(false))
-
-  const explorerTitle = bridgeNetwork?.blockExplorer.name || getChainInfo(chainId)?.explorerTitle || 'Explorer'
-
-  return (
-    <ConfirmDetailsItem
-      label={
-        <>
-          <TimelineIconCircleWrapper padding="0" bgColor={'transparent'}>
-            <StyledTimelineReceiptIcon src={ReceiptIcon} />
-          </TimelineIconCircleWrapper>{' '}
-          {label}
-        </>
-      }
-    >
-      <ExternalLink href={link}>{isMobile ? `${explorerTitle} ↗` : `View on ${explorerTitle} ↗`}</ExternalLink>
-    </ConfirmDetailsItem>
-  )
 }
 
 export function ReceivedBridgingContent({
@@ -63,7 +33,6 @@ export function ReceivedBridgingContent({
   const { data: bridgeSupportedNetworks } = useBridgeSupportedNetworks()
   const destinationBridgeNetwork = bridgeSupportedNetworks?.find((network) => network.id === destinationChainId)
 
-  const depositLink = depositTxHash && getExplorerLink(sourceChainId, depositTxHash, ExplorerDataType.TRANSACTION)
   const blockExplorerUrl = destinationBridgeNetwork?.blockExplorer?.url || getChainInfo(destinationChainId)?.explorer
 
   const fillTxLink =
@@ -85,8 +54,10 @@ export function ReceivedBridgingContent({
         </b>
       </ConfirmDetailsItem>
 
-      {depositLink && <TransactionLink link={depositLink} label="Source transaction" chainId={sourceChainId} />}
-      {fillTxLink && <TransactionLink link={fillTxLink} label="Destination transaction" chainId={destinationChainId} />}
+      <DepositTxLink depositTxHash={depositTxHash} sourceChainId={sourceChainId} />
+      {fillTxLink && (
+        <TransactionLinkItem link={fillTxLink} label="Destination transaction" chainId={destinationChainId} />
+      )}
     </>
   )
 }
